# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/comingtables/scan-barcode/{id}": {
            "post": {
                "description": "insert ComingTable",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SCAN_BARCODE"
                ],
                "summary": "Post From Scan",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ComingTable ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 1,
                        "description": "Count of Product",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by barcode",
                        "name": "barcode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    }
                }
            }
        },
...
        "/remaining": {
            "get": {
                "description": "get Remaining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "REMAINING"
                ],
                "summary": "List Remaining",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "limit for response",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page of req",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by categorId",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by branchId",
                        "name": "branch_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by barcode",
                        "name": "barcode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Remaining"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Remaining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "REMAINING"
                ],
                "summary": "Create Remaining",
                "parameters": [
                    {
                        "description": "Remaining data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUpdateRemaining"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    }
                }
            }
        },
        "/remaining/{id}": {
            "get": {
                "description": "get Remaining by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "REMAINING"
                ],
                "summary": "Get By Id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Remaining ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Remaining"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    }
                }
            },
            "put": {
                "description": "api for update Remaining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "REMAINING"
                ],
                "summary": "Update Remaining",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "id of Remaining",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data of Remaining",
                        "name": "remaining",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUpdateRemaining"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete Remaining by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "REMAINING"
                ],
                "summary": "Delete By Id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Remaining ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateUpdateRemaining": {
            "type": "object",
            "properties": {
                "barcode": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "count": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "models.Remaining": {
            "type": "object",
            "properties": {
                "barcode": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "count": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "total_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}